Title: 사각형 채우기 3, Time: 134ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 250513/사각형 채우기 3/rectangle-fill-3.java
@@ -5,7 +5,7 @@ public class Main {
 	
 	public static int n;
 	
-	public static int[] dp = new int[MAX_N+1];
+	public static long[] dp = new long[MAX_N+1];
 	
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
